requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Google Cloud and ADK dependencies
google-adk>=1.3.0
google-cloud-run>=0.10.0
google-cloud-build>=3.20.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-cloud-storage>=2.10.0

# DSPy for agent generation
dspy>=2.4.0

# vLLM for local model inference
vllm>=0.2.7
torch>=2.1.0
transformers>=4.36.0
tokenizers>=0.15.0

# Monitoring and observability
opik>=0.1.0

# Utility dependencies
numpy>=1.24.0
pandas>=2.0.0
pathlib-pathlib>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
